[API] Implement MatchLogView PUT,DELETE Method, Remove MatchDataView,MatchMemberView
