+++ API Server/Api/Views/MatchLogView.py
@@ -85,3 +85,62 @@ class MatchLogView(APIView):
                 }
 
         return JsonResponse(response_json)
+
+    def put(self, request):
+        MemberObj = CheckUserID(request)
+        tempdata = JSONParser().parse(request)
+        tempdata['mc_Driver'] = MemberObj.user_Id
+
+        if MemberObj.user_Driver:
+            match_serializer = MatchDataSerializer(data=tempdata)
+            if match_serializer.is_valid():
+                match_serializer.save()
+                response_json = {
+                    'success': True
+                }
+            else:
+                response_json = {
+                    'err': match_serializer.errors
+                }
+        else:
+            tempdata['mm_Member'] = MemberObj.user_Id
+            matchdata = MatchData.objects.get(mc_Driver=tempdata['mm_Driver'])
+
+            if matchdata:
+                tempdata['mm_Driver'] = matchdata.mc_Driver
+                matchmember_serializer = MatchMemberSerializer(data=tempdata)
+                if matchmember_serializer.is_valid():
+                    matchmember_serializer.save()
+                    response_json = {
+                        'success': True
+                    }
+                else:
+                    response_json = {
+                        'err': matchmember_serializer.errors
+                    }
+            else:
+                response_json = {
+                    'err': 'MatchData does not exist'
+                }
+
+        return JsonResponse(response_json)
+
+    def delete(self, request):
+        MemberObj = CheckUserID(request)
+        tempdata = JSONParser().parse(request)
+
+        if MemberObj.user_Driver:
+            #MatchData_id로 변경예정
+            MatchDataDel = MatchData.objects.get(mc_Driver=MemberObj)
+            MatchDataDel.delete()
+            response_json = {
+                'success': True
+            }
+        else:
+            #MatchMember_id or fk matchdata_id로 변경예정
+            MatchMemberDel = MatchMember.objects.get(mm_Member=MemberObj)
+            MatchMemberDel.delete()
+            response_json = {
+                'success': True
+            }
+        return JsonResponse(response_json)
